+++ libs/backroad-core/src/lib/core.ts
@@ -3,6 +3,7 @@ import type { Props } from 'react-select';
 import type { ColumnHelper } from '@tanstack/react-table';
 import { HTMLProps } from 'react';
 import formidable from 'formidable';
+import type { DropzoneOptions } from 'react-dropzone';
 // import { v4 as uuidv4 } from 'uuid';
 // type FileUploadObject = {
 //   id: string;
@@ -46,14 +47,18 @@ type _ComponentBasePropsMapping = {
     value: boolean;
   };
 
-  select: {
+  select: AllowDefaultHelper<{
     readonly args: {
       // options: any[];
       label?: string;
       // formatOption?: (option: any) => string;
-    } & Omit<Props<SelectOptionType, false>, 'onChange' | 'isMulti'>;
-    value: SelectValueType | null;
-  };
+    } & Omit<
+      Props<SelectOptionType, false>,
+      'onChange' | 'isMulti' | 'defaultValue'
+    >;
+
+    value: SelectValueType;
+  }>;
   multiselect: {
     args: {
       // options: any[];
@@ -172,10 +177,8 @@ type _ComponentBasePropsMapping = {
   }>;
   file_upload: {
     args: {
-      multiple?: boolean;
-      accept?: string;
       label: string;
-    };
+    } & Omit<DropzoneOptions, 'onUpload'>;
     value: formidable.File[];
   };
   video: {